+++ db/migrate/20151215202522_add_diagnosed_to_survey.rb
@@ -1,0 +1,5 @@
+class AddDiagnosedToSurvey < ActiveRecord::Migration
+  def change
+    add_column :surveys, :pediatric_diagnosed, :boolean, null: false, default: false
+  end
+end